Final Changes made, works as requested
The best score only changes if the new score is more than the previous score, also there was a bug that if the player changed their name, pressed start but exited the game before playing a game, the name on the menu would change but not the best score. The best score would still be the previous score and not 0. That bug has been fixed.